set owner reference on apiserver receive adapter
