+++ pkg/apis/sources/v1alpha1/apiserver_types.go
@@ -18,8 +18,10 @@ package v1alpha1
 
 import (
 	duckv1alpha1 "github.com/knative/pkg/apis/duck/v1alpha1"
+	"github.com/knative/pkg/kmeta"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
 // +genclient
@@ -35,6 +37,14 @@ type ApiServerSource struct {
 	Status ApiServerSourceStatus `json:"status,omitempty"`
 }
 
+// Check that we can create OwnerReferences to a Configuration.
+var _ kmeta.OwnerRefable = (*ApiServerSource)(nil)
+
+// GetGroupVersionKind returns the GroupVersionKind.
+func (s *ApiServerSource) GetGroupVersionKind() schema.GroupVersionKind {
+	return SchemeGroupVersion.WithKind("ApiServerSource")
+}
+
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
 
 // ApiServerSourceList contains a list of ApiServerSource
@@ -43,22 +53,6 @@ type ApiServerSourceList struct {
 	metav1.ListMeta `json:"metadata,omitempty"`
 	Items           []ApiServerSource `json:"items"`
 }
-
-const (
-	// ApiServerConditionReady has status True when the ApiServerSource is ready to send events.
-	ApiServerConditionReady = duckv1alpha1.ConditionReady
-
-	// ApiServerConditionSinkProvided has status True when the ApiServerSource has been configured with a sink target.
-	ApiServerConditionSinkProvided duckv1alpha1.ConditionType = "SinkProvided"
-
-	// ApiServerConditionDeployed has status True when the ApiServerSource has had it's deployment created.
-	ApiServerConditionDeployed duckv1alpha1.ConditionType = "Deployed"
-)
-
-var apiserverCondSet = duckv1alpha1.NewLivingConditionSet(
-	ApiServerConditionSinkProvided,
-	ApiServerConditionDeployed,
-)
 
 // ApiServerSourceSpec defines the desired state of ApiServerSource
 type ApiServerSourceSpec struct {
