+++ pkg/reconciler/apiserversource/resources/receive_adapter.go
@@ -20,6 +20,7 @@ import (
 	"fmt"
 
 	"github.com/knative/eventing/pkg/apis/sources/v1alpha1"
+	"github.com/knative/pkg/kmeta"
 	v1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -41,8 +42,11 @@ func MakeReceiveAdapter(args *ReceiveAdapterArgs) *v1.Deployment {
 	return &v1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace:    args.Source.Namespace,
-			GenerateName: fmt.Sprintf("apiserver-%s-", args.Source.Name),
+			GenerateName: fmt.Sprintf("apiserversource-%s-", args.Source.Name),
 			Labels:       args.Labels,
+			OwnerReferences: []metav1.OwnerReference{
+				*kmeta.NewControllerRef(args.Source),
+			},
 		},
 		Spec: v1.DeploymentSpec{
 			Selector: &metav1.LabelSelector{
